# Use fallocate when creating a new output