Using ftruncate() creates holes inside files, that on first (write) page fault gets allocated an actual disk block.  There are several disadvantages with this:
- Allocating block one-by-one is slower than doing it all at once.
- As we write to the output in parallel the write order is not exactly sequential.  The allocated blocks can also be out-of-order and make the on-disk representation fragmented.

Instead, use posix_fallocate() which allocates a contiguous on-disk block (if possible).  It's arguably a better fit for linker output, since holes are unlikely to be present after writing all the output (any contiguous zeroes should just go into .bss).

I benchmarked this by linking mold with debug symbols.  On an 8-core desktop, it is 5% (1.0s -> 0.95s) faster, but on a 64-core workstation it is actually 40% slower (0.37s -> 0.52s).  The total amount of System time is reduced (6.7s -> 5.2s), however.  I suspect that the reduced page fault work ends up causing more lock contention.

Despite potential regression on high-core systems I still think this change is justified because of the better on-disk layout.