@rui314 2023-12-05
The performance regression on the high core count machine is concerning as our focus is to speed up build time on such powerful machine. What if we consider an output file as 16 shards and call fallocate on each shard simultaneously from multiple threads?

Besides that, I believe we should use fallocate rather than posix_fallocate. As far as I know, posix_fallocate write to the entire file if fallocate is not supported by the underlying filesystem. If that happens, it'd be very slow.