@ishitatsuyuki 2023-12-05
Actually, it seems the slowdown is from some regression between 2.3.3 and 2.4.0 (I benched against the system mold which turns out to be outdated). The `User` column has increased significantly for `.debug_*` sections. I'll debug this issue separately.

Interesting note about the fallback behavior.  Since `posix_fallocate` seems to be absent for macOS anyway, I'll make it so that we try `fallocate` first if it exists and succeed, and otherwise call `ftruncate`.