@ishitatsuyuki 2023-12-05
Welp, after a clean rebuild I can't reproduce the slowness anymore. I probably messed up my build CFLAGS or something.

The good news is that performance is neutral after doing the benchmarking again (both around 0.36s, well within the margin of error).  I've removed the performance claims from the commit message.