@rui314 2023-12-05
I believe we still need to call `ftruncate` before or after `fallocate` in case the existing file is larger than the new size. I'll make the change after merging this commit.