@ishitatsuyuki 2023-12-05
The path for reusing an existing file is unchanged and still uses ftruncate directly.  But I guess using fallocate can make sense for it too.